refactor: route dns-record events through reconcile loop
Address review feedback: instead of dedicated handlers for dns-record
relation_created/joined that call _update_dns_records directly, route
these events through _on_config_changed -> _reconcile() like all other
relation events in this holistic charm.

Also migrates dns_record lib from charms.dns_record to
charms.dns_integrator namespace (dns-record charm doesn't exist on
Charmhub; dns-integrator does).
