+++ haproxy-operator/tests/unit/test_dns_record.py
@@ -6,7 +6,7 @@
 from unittest.mock import MagicMock
 
 import pytest
-from charms.dns_record.v0.dns_record import (
+from charms.dns_integrator.v0.dns_record import (
     CreateRecordRequestError,
     DNSRecordRequires,
     RecordRequest,
